backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
requests==2.32.3 